+++ flask_typed_routes/__init__.py
@@ -1,4 +1,4 @@
-from flask_typed_routes.app import FlaskTypeRoutes  # noqa
+from flask_typed_routes.app import FlaskTypedRoutes  # noqa
 from flask_typed_routes.errors import ValidationError  # noqa
 from flask_typed_routes.fields import Query, Path, Header, Cookie, JsonBody  # noqa
 from flask_typed_routes.core import typed_route  # noqa
@@ -11,5 +11,5 @@ __all__ = [
     "Cookie",
     "JsonBody",
     "typed_route",
-    "FlaskTypeRoutes",
+    "FlaskTypedRoutes",
 ]